Moves getLiveExampleState logic. Creates types.ts
