+++ components/pages/example/KnobRow/KnobRow.tsx
@@ -7,7 +7,7 @@ import { HTMLElementProps } from '@leafygreen-ui/lib';
 import InlineDefinition from '@leafygreen-ui/inline-definition';
 
 import { Knob } from '../Knob/Knob';
-import { KnobType } from '../utils';
+import { KnobType } from '../types';
 import { PropTooltipContent } from 'components/PropTooltipContent';
 
 const knobRowWrapperStyle = (darkMode: boolean) => css`
